Add axios for API calls and implement useProjects hook for project data management
